backend/requirements.txt
fastapi
uvicorn
openai
bcrypt
python-jose
python-dotenv
passlib
python-docx
boto3
sqlalchemy
python-multipart
celery[redis]
redis
python-socketio
eventlet
requests
beautifulsoup4
selenium
playwright
aiohttp
webdriver-manager
undetected-chromedriver
alembic
psycopg2-binary
pydantic[email]
websockets
